# apiFrotaMais/src/config/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Tecnologias Emergentes",
    "description": "Documentação da API criada em sala"
  },
  "servers": [
    {
      "url": "http://localhost:4040"
    }
  ],
  "paths": {
...
    "/api/maintenance/{id}": {
      "get": {
        "tags": [
          "Maintenance"
        ],
        "summary": "Obter manutenção por ID",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "_page",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "_order",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "_sort",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "Maintenance"
        ],
        "summary": "Atualizar manutenção existente",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "_page",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "_order",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "_sort",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Maintenance"
        ],
        "summary": "Deletar manutenção",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "_page",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "_order",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "_sort",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/api/phases/": {
      "get": {
        "tags": [
          "Maintenance Phases"
        ],
        "summary": "Listar fases de manutenções",
        "description": "",
        "parameters": [
          {
            "name": "_order",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "_sort",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Filtrar por status da manutenção (aprovado/concluido)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tipoFase",
            "in": "query",
            "description": "Filtrar por tipo de fase",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ativo",
            "in": "query",
            "description": "Filtrar por fases ativas (true/false)",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "_page",
            "in": "query",
            "description": "Número da página",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "_limit",
            "in": "query",
            "description": "Limite de itens por página",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/api/phases/active": {
      "get": {
        "tags": [
          "Maintenance Phases"
        ],
        "summary": "Listar fases de manutenções",
        "description": "",
        "parameters": [
          {
            "name": "_order",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "_sort",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Filtrar por status da manutenção (aprovado/concluido)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tipoFase",
            "in": "query",
            "description": "Filtrar por tipo de fase",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ativo",
            "in": "query",
            "description": "Filtrar por fases ativas (true/false)",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "_page",
            "in": "query",
            "description": "Número da página",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "_limit",
            "in": "query",
            "description": "Limite de itens por página",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/api/phases/iniciar-viagem": {
      "get": {
        "tags": [
          "Maintenance Phases"
        ],
        "summary": "Listar fases de manutenções",
        "description": "",
        "parameters": [
          {
            "name": "_order",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "_sort",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Filtrar por status da manutenção (aprovado/concluido)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tipoFase",
            "in": "query",
            "description": "Filtrar por tipo de fase",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ativo",
            "in": "query",
            "description": "Filtrar por fases ativas (true/false)",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "_page",
            "in": "query",
            "description": "Número da página",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "_limit",
            "in": "query",
            "description": "Limite de itens por página",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/api/phases/deixar-veiculo": {
      "get": {
        "tags": [
          "Maintenance Phases"
        ],
        "summary": "Listar fases de manutenções",
        "description": "",
        "parameters": [
          {
            "name": "_order",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "_sort",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Filtrar por status da manutenção (aprovado/concluido)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tipoFase",
            "in": "query",
            "description": "Filtrar por tipo de fase",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ativo",
            "in": "query",
            "description": "Filtrar por fases ativas (true/false)",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "_page",
            "in": "query",
            "description": "Número da página",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "_limit",
            "in": "query",
            "description": "Limite de itens por página",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/api/phases/servico-finalizado": {
      "get": {
        "tags": [
          "Maintenance Phases"
        ],
        "summary": "Listar fases de manutenções",
        "description": "",
        "parameters": [
          {
            "name": "_order",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "_sort",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Filtrar por status da manutenção (aprovado/concluido)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tipoFase",
            "in": "query",
            "description": "Filtrar por tipo de fase",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ativo",
            "in": "query",
            "description": "Filtrar por fases ativas (true/false)",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "_page",
            "in": "query",
            "description": "Número da página",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "_limit",
            "in": "query",
            "description": "Limite de itens por página",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/api/phases/retorno-veiculo": {
      "get": {
        "tags": [
          "Maintenance Phases"
        ],
        "summary": "Listar fases de manutenções",
        "description": "",
        "parameters": [
          {
            "name": "_order",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "_sort",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Filtrar por status da manutenção (aprovado/concluido)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tipoFase",
            "in": "query",
            "description": "Filtrar por tipo de fase",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ativo",
            "in": "query",
            "description": "Filtrar por fases ativas (true/false)",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "_page",
            "in": "query",
            "description": "Número da página",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "_limit",
            "in": "query",
            "description": "Limite de itens por página",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/api/phases/veiculo-entregue": {
      "get": {
        "tags": [
          "Maintenance Phases"
        ],
        "summary": "Listar fases de manutenções",
        "description": "",
        "parameters": [
          {
            "name": "_order",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "_sort",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Filtrar por status da manutenção (aprovado/concluido)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tipoFase",
            "in": "query",
            "description": "Filtrar por tipo de fase",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ativo",
            "in": "query",
            "description": "Filtrar por fases ativas (true/false)",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "_page",
            "in": "query",
            "description": "Número da página",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "_limit",
            "in": "query",
            "description": "Limite de itens por página",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/api/phases/approved": {
      "get": {
        "tags": [
          "Maintenance Phases"
        ],
        "summary": "Listar fases de manutenções",
        "description": "",
        "parameters": [
          {
            "name": "_order",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "_sort",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Filtrar por status da manutenção (aprovado/concluido)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tipoFase",
            "in": "query",
            "description": "Filtrar por tipo de fase",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ativo",
            "in": "query",
            "description": "Filtrar por fases ativas (true/false)",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "_page",
            "in": "query",
            "description": "Número da página",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "_limit",
            "in": "query",
            "description": "Limite de itens por página",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/api/phases/completed": {
      "get": {
        "tags": [
          "Maintenance Phases"
        ],
        "summary": "Listar fases de manutenções",
        "description": "",
        "parameters": [
          {
            "name": "_order",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "_sort",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Filtrar por status da manutenção (aprovado/concluido)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tipoFase",
            "in": "query",
            "description": "Filtrar por tipo de fase",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ativo",
            "in": "query",
            "description": "Filtrar por fases ativas (true/false)",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "_page",
            "in": "query",
            "description": "Número da página",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "_limit",
            "in": "query",
            "description": "Limite de itens por página",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/api/phases/maintenance/{manutencaoId}": {
      "get": {
        "tags": [
          "Maintenance Phases"
        ],
        "summary": "Obter todas as fases de uma manutenção",
        "description": "",
        "parameters": [
          {
            "name": "manutencaoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID da manutenção"
          },
          {
            "name": "_page",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "_order",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "_sort",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Fases da manutenção"
          },
          "403": {
            "description": "Acesso proibido"
          },
          "404": {
            "description": "Manutenção não encontrada"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/api/phases/maintenance/{manutencaoId}/phase": {
      "post": {
        "tags": [
          "Maintenance Phases"
        ],
        "summary": "Criar nova fase de manutenção",
        "description": "",
        "parameters": [
          {
            "name": "manutencaoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "_page",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "_order",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "_sort",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "in": {
                    "example": "any"
                  },
                  "description": {
                    "example": "any"
                  },
                  "required": {
                    "example": "any"
                  },
                  "schema": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/phases/maintenance/{manutencaoId}/advance": {
      "post": {
        "tags": [
          "Maintenance Phases"
        ],
        "summary": "Avançar para próxima fase da manutenção ou finalizar última fase",
        "description": "",
        "parameters": [
          {
            "name": "manutencaoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID da manutenção"
          },
          {
            "name": "_page",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "_order",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "_sort",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "in": {
                    "example": "any"
                  },
                  "description": {
                    "example": "any"
                  },
                  "required": {
                    "example": "any"
                  },
                  "schema": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/phases/{id}": {
      "get": {
        "tags": [
          "Maintenance Phases"
        ],
        "summary": "Obter fase de manutenção por ID",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID da fase de manutenção"
          },
          {
            "name": "_page",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "_order",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "_sort",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detalhes da fase de manutenção"
          },
          "403": {
            "description": "Acesso proibido"
          },
          "404": {
            "description": "Fase de manutenção não encontrada"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "Maintenance Phases"
        ],
        "summary": "Atualizar fase de manutenção",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID da fase de manutenção"
          },
          {
            "name": "_page",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "_order",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "_sort",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "in": {
                    "example": "any"
                  },
                  "description": {
                    "example": "any"
                  },
                  "required": {
                    "example": "any"
                  },
                  "schema": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Insira o token JWT"
      }
    },
...
  }
}